==> 1.1.OrganizeWVAdata.Rmd <==
---
title: "West Virginia Data Organization"
author: "Emma Jones"
date: "8/9/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
library(tidyverse)
library(sf)
#library(DBI)
library(lubridate)
#library(pool)
#library(pins)
#library(config)
library(readxl)
#library(dbplyr)

```

## Background

This document walks through the data organization process of bringing in select WVA sites into the IBI development process to flesh out some areas where VA lacks sufficient data.

### Bugs

First bring in the WVA stations we want to use for IBI work (reference sites in central apps) and spring and fall final bug information from the Virginia Tolerance Value Attribution work circa 2018.

```{r bug tol work}
WVA <- read_excel('data/Jason/RefStressDataFinal_WV_Final_Emma.xlsx', sheet = 'JRH_Final_WVDEP-LatlongEco') %>% 
  filter(JRH_Code == 'Ref')


allSpringBugs <- read_csv('C:/HardDriveBackup/R/GitHub/VAattribution_round2/data/processedData/allSpringBugs.csv') %>% 
  mutate(Season = 'Spring') %>% 
  dplyr::select(UID, DataSource, StationID, CollDate, Year, Season, everything())
allFallBugs <- read_csv('C:/HardDriveBackup/R/GitHub/VAattribution_round2/data/processedData/allFallBugs.csv') %>% 
  mutate(Season = 'Fall') %>% 
  dplyr::select(UID, DataSource, StationID, CollDate, Year, Season, everything())
```

Keep only the bug records from sites we want.

```{r WVA slim}
WVAbugs <- bind_rows(allSpringBugs, allFallBugs) %>% 
  filter(StationID %in% WVA$JRH_Final_WVDEP.StationID) %>% 
  # first group by StationID and Date to drop duplicate Rep1 samples
  group_by(StationID, CollDate) %>% 
  mutate(SampleN = 1:n()) %>% 
  # look at samples with multiple rep1's
  # dplyr::select(SampleN, everything()) %>%  
  # arrange(desc(SampleN))
  # write.csv(benSamps1, 'multiple rep 1s.csv', row.names = F)
  filter(! SampleN > 1) %>%  # drop multiple rep 1s before filtering on max 4 samples per site
  ungroup() %>% 
  
  # then drop multiple reps of the same season (not taken on the same day)
  ##mutate(Year = year(`Collection Date`)) %>% 
  group_by(StationID, Year, Season) %>% 
  mutate(SampleN = 1:n()) %>% 
  # look at samples with multiple rep1's in a season
  # dplyr::select(SampleN, Year, Season, everything()) %>% 
  # arrange(desc(SampleN)) %>% 
  filter(! SampleN > 1) %>% 
  
  # then total all samples taken and drop any records > 4 for a station, starting the count at the most recent sample event
  ungroup() %>% 
  group_by(StationID) %>% 
  arrange(desc(CollDate)) %>% 
  mutate(CollDate = as.Date(CollDate),
         SampleN = 1:n(),# number samples from most recent collection
         JulianDate = yday(CollDate)) %>%  
  # dplyr::select(SampleN, Year, Season, everything()) %>% 
  # arrange(desc(SampleN)) %>% 
  filter(SampleN <= 4)  # only keep most recent four samples


# Any ref sites with no data?
filter(WVAbugs, ! StationID %in% WVA$JRH_Final_WVDEP.StationID) # no rows, cool.
```


## Grab any environmental data available from Tol Value project

```{r tol value env data}
WVAenvData <- read_csv('C:/HardDriveBackup/R/GitHub/VAattribution_round2/data/processedData/finalEnvData.csv') %>% 
  filter(UID %in% WVAbugs$UID)

# any of this info from the original spreadsheet?
filter(WVAenvData, is.na(Order)) %>% 
  filter(StationID %in% filter(WVA, !is.na(Order))$JRH_Final_WVDEP.StationID)
# nope. Need to get from WVA

# write.csv(
# filter(WVAenvData, is.na(Order)) %>% 
#   distinct(StationID, .keep_all = T) %>% 
#   dplyr::select(DataSource, StationID, LongitudeDD, LatitudeDD, Order), 'data/needSpatial/MissingOrder_WVA.csv', row.names = F)
```

Bring in order information from WVA. Order based on 1:100k NHD

```{r WVA order}
WVAorderData <- read_csv('data/needSpatial/MissingOrder_WVA_fromWVA.csv') %>% 
  # apply same logic as before to streams not on 1:100k NHD
  mutate(Order = case_when(Order == 0 ~ 1, 
                           TRUE ~ as.numeric(Order)))

WVAenvData <- left_join(WVAenvData, WVAorderData, by = c('DataSource', 'StationID', 'LongitudeDD', 'LatitudeDD')) %>% 
  mutate(Order = coalesce(Order.x, Order.y)) %>% 
  dplyr::select(DataSource:BioPanel, Order, everything()) %>% 
  dplyr::select(-c(Order.x, Order.y))

filter(WVAenvData, is.na(Order)) # okay bc just fish sites and fish will use area not order
```


Finish matching env data format to VA data.


```{r env data smash}
envData <- read.csv('data/dataOut/stationsSpatial_withAllData.csv') %>% 
  rename('Collection Date' = 'Collection.Date') %>% 
  mutate(`Collection Date` = as.Date(as.character(`Collection Date`)) )

WVAenvDataFin <- WVAenvData %>%
  left_join(dplyr::select(WVAbugs, UID, StationID, `Collection Date` = CollDate, JulianDate),
            by = c('UID', 'StationID')) %>% 
  mutate(JRH_Final_Ref_Cod = as.character(NA),
         Coastal = as.logical(NA),
         monthday = as.numeric(paste0(sprintf("%02d",month(`Collection Date`)),
                                      sprintf("%02d",day(`Collection Date`)))),
         Season = case_when(monthday >= 0215 & monthday <= 0615 ~ 'Spring',
                            monthday >= 0815 & monthday <= 1215 ~ 'Fall',
                            TRUE ~ as.character("Outside Sample Window")),
         totalArea_sqMile = AREA_SQ_MILES) %>% 
  rename("Latitude" = "LatitudeDD", "Longitude" = "LongitudeDD", "US_L3NAME"= "EcoRegion") %>% 
  dplyr::select(DataSource, JRH_Final_Ref_Cod , Coastal, StationID, UID, Year, Season, JulianDate, `Collection Date`,
                Latitude, Longitude, US_L3NAME, totalArea_sqMile, Order)

envDataVA_WVA <- bind_rows(envData, WVAenvDataFin)

```



Clean up workspace

```{r}
rm( allFallBugs); rm(allSpringBugs); rm(WVAorderData)
```


## Compare results to DEQ bug info 

Bring in DEQ's bug data to identify any WVA taxonomic calls that may need to be adjusted to fit DEQ's data model.

Make a list of DEQ's bugs

```{r DEQ bugTime}
benthics <- read_csv('data/dataOut/benthics_ExcludedTaxaDropped.csv') %>% 
  pivot_longer(cols = -c(StationID:RepNum), names_to = 'Taxa', values_to = 'Individuals') %>% 
  distinct(Taxa) %>% 
  #dplyr::select(DEQtaxa = Taxa) %>% 
  mutate(Agency = 'VA DEQ')
```

Same for WVA. 

```{r real WVA work}
benthicsWVA <- WVAbugs %>% ungroup() %>% 
  pivot_longer(cols = -c(UID:Season, SampleN, JulianDate, RepNum), names_to = 'Taxa', values_to = 'Individuals') %>% 
  distinct(Taxa) %>% 
  mutate(Agency = 'WV DEP')
  #dplyr::select(WVAtaxa = Taxa)
```

Smash together to figure out what is in WVA and not in DEQ

```{r bug smash}
taxa <- full_join(benthics, benthicsWVA, by = 'Taxa')

NeedHelp <- filter(taxa, is.na(Agency.x))
```

What taxa are valid (in VA's master taxa list but not in VA's ref site database) and what need additional work to use for this project?

```{r master Taxa}
library(pins)

# get ODS and server configuration settings
conn <- config::get("connectionSettings")

# connect to server
board_register_rsconnect(key = conn$CONNECT_API_KEY,  #Sys.getenv("CONNECT_API_KEY"),
                         server = conn$CONNECT_SERVER)#Sys.getenv("CONNECT_SERVER"))

masterTaxa <- pin_get('ejones/masterTaxaGenus', board = 'rsconnect')


NeedHelp1 <- filter(taxa, is.na(Agency.x)) %>% 
  dplyr::select(-contains('Agency')) %>% 
  filter(! Taxa %in% masterTaxa$FinalID) %>% 
  left_join(masterTaxa %>% mutate(Genus1 = Genus), by = c("Taxa" = 'Genus1')) %>% # make extra Genus column to make the master taxa look intact
  arrange(Genus)

```


ask for help.

```{r}
# write_csv(taxa, 'data/dataOut/VA_WVA taxa join.csv', na="")
# write_csv(NeedHelp1, 'data/dataOut/WVA taxa help.csv', na="")
```


Greg Pond help file.

```{r GregPond help}
pond <- read_excel('data/Jason/WVA taxa help_GPcomments.xlsx')

acceptGenus <- filter(pond, !is.na(FinalID)) %>% 
  dplyr::select(Taxa, FinalID)

benthicsWVAfin <- WVAbugs %>% ungroup() %>% 
  pivot_longer(cols = -c(UID:Season, SampleN, JulianDate, RepNum), names_to = 'Taxa', values_to = 'Individuals') %>% 
  filter(Individuals > 0 ) %>% 
  left_join(acceptGenus, by = 'Taxa') %>%  # join in Greg's easy calls
  mutate(Taxa = case_when(!is.na(FinalID) ~ FinalID,
                          TRUE ~ Taxa)) %>% # recode FinalID to Taxa name where appropriate
  dplyr::select(-FinalID) %>% 
  # map the more difficult ones by hand
  mutate(Taxa = case_when(Taxa == 'Ansiocentropus' ~ 'Anisocentropus',
                          Taxa == 'Astioplax dolani' ~ 'Asioplax dolani',
                          Taxa == 'Chironomidae (A)' ~ 'Chironomidae',
                          Taxa == 'Libellulidae/Cordullidae' ~ 'Corduliidae',
                          Taxa == 'Lymnaidae' ~ 'Lymnaeidae',
                          Taxa == 'Nematomrpha' ~ 'Nematomorpha',
                          Taxa == 'Nematopmorpha' ~ 'Nematomorpha',
                          Taxa == 'Phyrganaeidae' ~ 'Phryganaeidae',
                          Taxa == 'Phyrganea' ~ 'Phryganea',
                          Taxa == 'Potmayia flava' ~ 'Potamayia flava',
                          Taxa == 'Sphaeriidae' ~ 'Pisidiidae',
                          Taxa == 'Stratiomyiidae' ~ 'Stratiomyidae',
                          TRUE ~ Taxa)) %>% 
  filter(! Taxa %in% c('Americaenis', 'Glossophoniidae','Helobdellidae', 'Nematomorpha', 'Nematomorpha', 
                        'Phryganaeidae', 'Potamayia flava', 'Pyralidae','Neopleiidae'))# %>% 
  # group_by(UID, DataSource, StationID, CollDate, Year, Season, RepNum, SampleN, JulianDate, Taxa) %>% 
  # mutate(n = n())
  # 
  # summarise(Individuals = sum(Individuals, na.rm = T))
  
# run taxa analysis again to make sure everything matches up
benthicsWVAfinQA <- benthicsWVAfin  %>%
  distinct(Taxa) %>% 
  mutate(Agency = 'WV DEP')

taxa <- full_join(benthics, benthicsWVAfinQA, by = 'Taxa')

# bring in VA again
benthics <- read_csv('data/dataOut/benthics_ExcludedTaxaDropped.csv') %>% 
  left_join(dplyr::select(envData, StationID, UID, `Collection Date`, DataSource), 
            by = c('StationID', 'Collection Date')) %>% 
  dplyr::select(DataSource, UID, everything())

#pivot wider, smash with VA, double check everything fits
benthicsWVAfinWide <- benthicsWVAfin %>% 
  pivot_wider(id_cols = c("UID", "DataSource","StationID", "CollDate", "Year", "Season", "RepNum", "SampleN", "JulianDate"),
              names_from = 'Taxa', values_from = 'Individuals') %>% 
  mutate(BenSampID = as.character(NA),
         `Collection DateTime` = as.POSIXct(NA)) %>% 
  dplyr::select(DataSource, UID, StationID, BenSampID, `Collection Date` = CollDate,  `Collection DateTime`, RepNum, everything())

benthicsVA_WVA <- bind_rows(
  benthics,
  benthicsWVAfinWide)
```


Save work

```{r save combined}
write_csv(benthicsVA_WVA, 'data/dataOut/benthicsVA_WVA.csv', na="")
write_csv(envDataVA_WVA, 'data/dataOut/envDataVA_WVA.csv', na="")

```
