==> 4.aquametBioBugMetrics.Rmd <==
---
title: "aquametBio Bug Metrics"
author: "Emma Jones"
date: "1/7/2022"
output: html_document
---

# Run in R 3.6.2

Run this part in 3.6.2 to use pins and pool (issue in 3.6.1) and preserve data for analysis in R 4.1.2. 

This will be better long term to have archive of data instead of pulling from server for analyses.

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
# 
# library(tidyverse)
# library(sf)
# library(pool)
# library(DBI)
# library(pins)
# library(config)
# library(lubridate)
# library(dbplyr)
# 
# 
# # get configuration settings
# conn <- config::get("connectionSettings")
# 
# # use API key to register board
# board_register_rsconnect(key = conn$CONNECT_API_KEY,  #Sys.getenv("CONNECT_API_KEY"),
#                           server = conn$CONNECT_SERVER)#Sys.getenv("CONNECT_SERVER"))
# 
# ## For testing: connect to ODS production
# pool <- dbPool(
#   drv = odbc::odbc(),
#   Driver = "ODBC Driver 11 for SQL Server",#Driver = "SQL Server Native Client 11.0",
#   Server= "DEQ-SQLODS-PROD,50000",
#   dbname = "ODS",
#   trusted_connection = "yes"
# )
```

## Background

This script picks up a few months after initial data organization and reference site analyses. This script pulls all bethic data and runs the 125 metrics provided in the aquametBio package. 

### Pull benthic data

```{r}
# benthics <- pool %>% tbl(in_schema("wqm",  "Edas_Benthic_View")) %>%
#   filter(WBS_TARGET_COUNT == 200) %>%
#         as_tibble() %>%
#         rename( "StationID" = "STA_ID",
#                 "BenSampID"  = "WBS_SAMP_ID",
#                 "RepNum" = "WBS_REP_NUM",
#                 "FinalID" = "WBMT_FINAL_ID",
#                 "Individuals" = "WBE_INDIVIDUALS",
#                 "ID Comments" = "WBE_COMMENT",
#                 "Entered By" = "WBE_INSERTED_BY", # not in EDAS table but good info
#                 "Taxonomist" = "TAXONOMIST_NAME",  # not in EDAS table but good info
#                 "Entered Date" = "WBE_INSERTED_DATE") %>%
#         mutate(`Excluded Taxa` = ifelse(WBE_EXCLUDED_TAXA_YN == "Y", -1, 0)) %>%
#         dplyr::select(StationID, BenSampID, RepNum, FinalID, Individuals, `Excluded Taxa`, `ID Comments`, Taxonomist, `Entered By`, `Entered Date`)
# 
# benSamps <- pool %>% tbl(in_schema("wqm",  "Edas_Benthic_Sample_View")) %>%
#   filter(WBS_SAMP_ID %in% !! unique(benthics$BenSampID)) %>%
#         as_tibble() %>%
#         # fix names
#         rename( "StationID" = "STA_ID",
#                 "BenSampID"  = "WBS_SAMP_ID",
#                 "RepNum" = "WBS_REP_NUM",
#                 "Sample Comments" = "WBS_COMMENT",
#                 "Entered By" = "WBS_INSERTED_BY", # not in EDAS table but good info
#                 "Collected By" = "COLLECTOR_NAME",  # not in EDAS table but good info
#                 "Entered Date" = "WBS_INSERTED_DATE",
#                 "Gradient" = "WBCM_DESCRIPTION",
#                 "Taxonomist" = "TAXONOMIST_NAME",  # not in EDAS table but good info
#                 "Target Count" = "WBS_TARGET_COUNT",
#                 "Field Team" = "WBS_FIELD_TEAM",
#                 "Collection Date" = "FDT_DATE_TIME") %>%
#         # Add sample season
#         mutate(monthday = as.numeric(paste0(sprintf("%02d",month(`Collection Date`)),
#                                             sprintf("%02d",day(`Collection Date`)))),
#                Season = case_when(monthday >= 0215 & monthday <= 0615 ~ 'Spring',
#                                   monthday >= 0815 & monthday <= 1215 ~ 'Fall',
#                                   TRUE ~ as.character("Outside Sample Window"))) %>%
#         dplyr::select(StationID, BenSampID, RepNum, `Collection Date`, `Sample Comments`, `Collected By`, `Field Team`, `Entered By`,
#                       Taxonomist, `Entered Date`, Gradient, `Target Count`, Season)
# 
# write.csv(benthics, 'data/aquametBioDataOut/benthics01132022.csv', row.names = F, na = '')
# write.csv(benSamps, 'data/aquametBioDataOut/benSamps01132022.csv', row.names = F, na = '')

```




# Run in 4.1.2
Run in R 4.1.2 for aquametBio and Hmisc

```{r}
library(tidyverse)
library(aquametBio)
library(Hmisc) # need to get %nin% to work
library(readxl)
source('aquametBio_specialFunctions/function_calcBentTaxMetsEVJ.R')
source('aquametBio_specialFunctions/function_calcAllBentMetsEVJ.R')
source('aquametBio_specialFunctions/function_prepBentCts_WSAEVJ.R')

```


Bring in master taxa list with genus calls

```{r}
masterTaxaList <- read_excel('data/masterTaxaGenus_01132022_JRH.xlsx',  sheet = 'masterTaxaGenusJRHFeb10') %>%  #change with genus fixes
  #read_excel('data/masterTaxaGenus_01132022.xlsx', sheet = 'masterTaxaGenus') %>% # Emma update
  mutate_at(c("FamTolVal", "TolVal", "GVSCI_TolVal"), as.numeric) %>% 
  rename(FinalID_FFG = FFG) %>% # fix this before things get confusing
  
  # Optioservus fix
  mutate(Genus = case_when(FinalID == 'Promoresia' ~ 'Promoresia',
                           TRUE ~ as.character(Genus))) %>% 
  # Stenonema fix
  mutate(Genus = case_when(FinalID %in% c("Maccaffertium/Stenonema", "Maccaffertium") ~ "Maccaffertium",
                           TRUE ~ as.character(Genus)),
         GVSCI_TolVal = case_when(FinalID %in% c("Maccaffertium/Stenonema", "Maccaffertium") ~ 5.4,
                           TRUE ~ as.numeric(GVSCI_TolVal)) ) %>% #,
         # GregVSCI_TolVal2 = case_when(FinalID %in% c("Maccaffertium/Stenonema", "Maccaffertium") ~ 5.4,
         #                   TRUE ~ as.numeric(GregVSCI_TolVal2))) %>% 
  
  
  # rename to desired format
  rename(PHYLUM = Phylum, 
         CLASS = Class,
         ORDER = Order, 
         FAMILY = Family,
         SUBFAMILY = Subfamily, 
         TRIBE = Tribe, 
         GENUS = Genus,
         # use genus level bugs for metrics
         #TARGET_TAXON = GVSCI_FinalID,
         FFG = GVSCI_FFG,
         HABIT = GVSCI_Habit, 
         PTV = GVSCI_TolVal) %>% 
  #mutate(TAXA_ID = 1:n()) %>%  # give numeric unique identifier
  mutate_if(is.character, str_to_upper) %>% # package needs uppercase taxa names
  
  # recode FFG to desired format: CF=collector-filterer, CG=collector-gatherer,PA=parasite,PI=Piercer,PR=predator,SC=scraper,SH=shredder
  mutate(FFG = case_when(FFG == "COLLECTOR" ~ "CG", 
                         FFG == "SCRAPER" ~ "SC",
                         FFG == "SHREDDER" ~ "SH",
                         FFG == "PREDATOR" ~ "PR",
                         FFG == "FILTERER" ~ "CF",
                         TRUE ~ as.character(NA)),
 #  recode HABIT to desired format:  AT=attached,BU=burrower,CB=climber, CN=clinger,DV=diver,PK=planktonic,SK=skater,SP=sprawler,SW=swimmer    
         HABIT = case_when(HABIT == "SWIMMER" ~ "SW", 
                           HABIT == "CLINGER" ~ "CN",
                           HABIT ==  "CLIMBER" ~ "CB",
                           HABIT == "BURROWER" ~ "BU",
                           HABIT == "SPRAWLER" ~ "SP",
                           TRUE ~ as.character(NA))) 
    
masterTaxaList[ masterTaxaList == "NA" ] <- NA

masterTaxaListTarget <- masterTaxaList %>% 
  #filter(Target == 1) # this will drop all higher level taxa from joining, not a good idea
  arrange(GVSCI_FinalID, Target) %>% 
  
    filter(!is.na(GVSCI_FinalID)) %>% 

  
  distinct(GVSCI_FinalID, .keep_all = T) %>% 
  rename(TARGET_TAXON = GVSCI_FinalID) %>% 
  mutate(TAXA_ID = 1:n()) %>% 
  as.data.frame()
  #rowwise() %>% 
  #mutate(TAXA_ID = sum(3000, TAXA_ID))
```
Bring in archived benthics info

```{r}
benthics <- read.csv('data/aquametBioDataOut/benthics01132022.csv') 
benSamps <- read.csv('data/aquametBioDataOut/benSamps01132022.csv') %>% 
  rename(`Collection Date` = 'Collection.Date',
         "Sample Comments" = "Sample.Comments", 
         "Collected By" = "Collected.By",
         "Target Count" = "Target.Count") %>% 
  mutate(DataSource = 'VDEQ_BUG',
         `Collection Date` = as.Date(as.POSIXct(`Collection Date`, format = '%Y-%m-%d %H:%M:%S'))) 
```


Bring in WVA bugs

```{r WVA bugs}
# see WVAbugmetrics.R to verify bug QA
benSampsWVA <- read_csv('data/dataOut/benthicsVA_WVA.csv') %>% 
  filter(DataSource == 'WVA_BUG') %>% 
    mutate(BenSampID = UID) %>%
  dplyr::select(DataSource, StationID, BenSampID, `Collection Date`, RepNum, Season )

benthicsWVA <- read_csv('data/dataOut/benthicsVA_WVA.csv') %>% 
  filter(DataSource == 'WVA_BUG') %>% 
  mutate(BenSampID = UID) %>% 
  dplyr::select(-c(UID, `Collection DateTime`, RepNum, Season, Year, JulianDate, SampleN)) %>% 
  group_by(DataSource, StationID, BenSampID, `Collection Date`) %>% 
  pivot_longer(-c('DataSource', 'StationID', 'BenSampID', 'Collection Date'), names_to = 'FinalID', values_to = 'Individuals') %>% 
  filter(Individuals > 0) %>% 
  #rename(BENSAMPID = BenSampID) %>%  # assignDistinct() changes names to caps so adjust here to avoid issues later
  #mutate(FinalID = as.character(FinalID)) %>% 
  #mutate_if(is.character, str_to_upper) %>%    # convert names to all caps
  # convert names to newest taxa list
  mutate(FinalID = case_when(FinalID == 'Nixe' ~ 'Afghanurus',
                             FinalID == 'Anchytarsus' ~ 'Anchytarsus bicolor', 
                             FinalID == 'Stenonema' ~ 'Stenonema femoratum',
                             TRUE ~ FinalID))

```

Smash VA and WVA together 
```{r smash VA WVA bugs}
benthicsAll <- bind_rows(benthics, benthicsWVA) %>% 
  dplyr::select(-c(DataSource, `Collection Date`)) %>% 
  group_by(BenSampID) %>% 
  mutate(TotalIndividuals = sum(Individuals)) %>% 
  ungroup()

benSamps <- bind_rows(benSamps, benSampsWVA)
```


# Rarify samples

If a sample has > 220 bugs then subsample back to 200ct
Repeat this only once as to not change analysis results.

```{r rarify}
# source("rarify/rarify.r")
# benthicsFine <- filter(benthicsAll, TotalIndividuals <= 220)
# benthicsFix <- filter(benthicsAll, TotalIndividuals > 220) 
# 
# rarified <- rarify(inbug=data.frame(benthicsFix), sample.ID="BenSampID" , abund="Individuals", subsiz=200) %>% 
#   filter(Individuals > 0) %>% 
#   group_by(BenSampID) %>% 
#   mutate(BenSampID = paste0(BenSampID, '_R200'),
#     TotalIndividuals = sum(Individuals))
# 
# benthics <- bind_rows(benthicsFine, rarified)
# 
# # fix benSamps with new BenSampID's
# benSampsFine <- filter(benSamps, BenSampID %in% benthicsFine$BenSampID)
# benSampsFix <- filter(benSamps, BenSampID %in% benthicsFix$BenSampID) %>% 
#    mutate(BenSampID = paste0(BenSampID, '_R200'))
# 
# benSamps <- bind_rows(benSampsFine, benSampsFix)
# 
# # write.csv(benthics, 'data/aquametBioDataOut/finalBenthics.csv', row.names = F, na = '')
# # write.csv(benSamps, 'data/aquametBioDataOut/finalBenSamps.csv', row.names = F, na = '')
# 
# 
# rm(benthicsAll);rm(benthicsFine);rm(benthicsFix);rm(benthicsWVA); rm(benSampsWVA); rm(benSampsFine); rm(benSampsFix);rm(benthics); rm(benSamps)
# rm(rarified)
```


Read in final benthics, bensamps

```{r benthics bensamps}
benthics <- read_csv('data/aquametBioDataOut/finalBenthics.csv')
benSamps <- read_csv('data/aquametBioDataOut/finalBenSamps.csv')
```





Clean bug data for aquametBio analysis

```{r}
benthics <- benthics %>% #[1:6921,] %>% 
  
# filter(BenSampID == 'CUB7141') %>% #'3-RAP10109') %>% 
  
  rename(BENSAMPID = BenSampID) %>%  # assignDistinct() changes names to caps so adjust here to avoid issues later
  mutate(FinalID = as.character(FinalID)) %>% 
  mutate_if(is.character, str_to_upper) %>%   # convert names to all caps
  
  ## join to full master taxa list
  #left_join(dplyr::select(masterTaxaList, FinalID, TARGET_TAXON, TAXA_ID),  by = "FinalID")
  left_join(dplyr::select(masterTaxaList, FinalID, GVSCI_FinalID#TARGET_TAXON),  by = "FinalID")
  ),  by = "FinalID") %>% 
  # join to target level master taxa list for unique TAXA_ID
  left_join(dplyr::select(masterTaxaListTarget, TARGET_TAXON, TAXA_ID),  by = c("GVSCI_FinalID" = 'TARGET_TAXON')) #"FinalID"
 # left_join(dplyr::select(masterTaxaListTarget, FinalID, TARGET_TAXON, TAXA_ID),  by = "FinalID")
  

  
benthicsPrep <- prepBentCts_WSA_EVJ(
  inCts = benthics,
  inTaxa = masterTaxaListTarget,
  sampID = "BENSAMPID",
  ct = "Individuals",
  taxa_id ="TAXA_ID"# "FinalID"#"TAXA_ID"
)

# EPA version recodes some bugs to their TAXA_ID and not ours
# benthicsPrep <- prepBentCts_WSA(
#   inCts = benthics,
#   inTaxa = masterTaxaListTarget,
#   sampID = "BENSAMPID",
#   ct = "Individuals",
#   taxa_id ="TAXA_ID"# "FinalID"#"TAXA_ID"
# )

```


QA benthicsPrep

```{r}
benthicsPrepReview <- left_join(benthicsPrep, masterTaxaListTarget) %>% 
  dplyr::select(BENSAMPID, TAXA_ID, TOTAL, IS_DISTINCT,
                TARGET_TAXON, FinalID, PTV, FFG, HABIT, everything())
```

Split benthicsPrep into 4 objects for taxonomic metric calculation (adding Total taxa/individuals destroyed function efficiency)

```{r split by 4}
uniqueBenthics <- benthicsPrep %>% 
  distinct(BENSAMPID) %>% 
  mutate(uniqueBenSampID = 1:n())

benthicsPrep1 <- left_join(benthicsPrep, uniqueBenthics, by = 'BENSAMPID') %>% 
  filter(uniqueBenSampID %in% c(1:1691)) %>% 
  dplyr::select(-uniqueBenSampID)
benthicsPrep2 <- left_join(benthicsPrep, uniqueBenthics, by = 'BENSAMPID') %>% 
  filter(uniqueBenSampID %in% c(1692:3382)) %>% 
  dplyr::select(-uniqueBenSampID)
benthicsPrep3 <- left_join(benthicsPrep, uniqueBenthics, by = 'BENSAMPID') %>% 
  filter(uniqueBenSampID %in% c(3383:5073)) %>% 
  dplyr::select(-uniqueBenSampID)
benthicsPrep4 <- left_join(benthicsPrep, uniqueBenthics, by = 'BENSAMPID') %>% 
  filter(uniqueBenSampID %in% c(5074:nrow(uniqueBenthics))) %>% 
  dplyr::select(-uniqueBenSampID)

```


Run metrics

```{r}
# testMetrics <- calcAllBentMetsEVJ(indf = benthicsPrep,
#                                inTaxa = masterTaxaListTarget,#masterTaxaList %>% mutate(TAXA_ID = FinalID), 
#                                sampID = "BENSAMPID", #c('UID','SAMPLE_TYPE','SAMPLE_CAT') option to concat more
#                                dist = "IS_DISTINCT", 
#                                ct = "TOTAL", 
#                                taxa_id = "TAXA_ID", 
#                                ffg = "FFG",#"FFG", 
#                                habit = "HABIT",#"Habit", 
#                                ptv = "PTV")#"TolVal")
```

Tack on benthic sampling info

```{r}
# testMetrics1 <- left_join(testMetrics, benSamps, by = c("BENSAMPID" = 'BenSampID')) %>% 
#   dplyr::select(StationID, BenSampID = BENSAMPID, RepNum, `Collection Date`, `Sample Comments`,
#                 `Collected By`, Taxonomist, Gradient, `Target Count`, Season, everything()) %>% 
#   arrange(`Collection Date`, StationID, RepNum)
# write.csv(testMetrics1, 'data/aquametBioDataOut/aquametBio_run110.csv', row.names = F, na = '')
```

Run groups of metrics and then save as individual sheets for IBI work

Taxa metrics
```{r taxa metrics}
# outTax <- calcBentTaxMetsEVJ(benthicsPrep,
#                           masterTaxaListTarget,
#                           sampID = "BENSAMPID", 
#                           dist='IS_DISTINCT',
#                           ct='TOTAL')



outTax1 <- calcBentTaxMetsEVJ(benthicsPrep1,
                          masterTaxaListTarget,
                          sampID = "BENSAMPID", 
                          dist='IS_DISTINCT',
                          ct='TOTAL')
outTax2 <- calcBentTaxMetsEVJ(benthicsPrep2,
                          masterTaxaListTarget,
                          sampID = "BENSAMPID", 
                          dist='IS_DISTINCT',
                          ct='TOTAL')
outTax3 <- calcBentTaxMetsEVJ(benthicsPrep3,
                          masterTaxaListTarget,
                          sampID = "BENSAMPID", 
                          dist='IS_DISTINCT',
                          ct='TOTAL')
outTax4 <- calcBentTaxMetsEVJ(benthicsPrep4,
                          masterTaxaListTarget,
                          sampID = "BENSAMPID", 
                          dist='IS_DISTINCT',
                          ct='TOTAL')


outTax <- bind_rows(outTax1, outTax2, outTax3, outTax4)
# clean up workspace
rm(benthicsPrep1);rm(benthicsPrep2);rm(benthicsPrep3);rm(benthicsPrep4)
rm(outTax1);rm(outTax2);rm(outTax3);rm(outTax4); rm(uniqueBenthics)
```

Tolerance metrics

```{r}
outTol <- calcBentTolMets(benthicsPrep,
                          masterTaxaListTarget,
                          sampID="BENSAMPID",
                          dist='IS_DISTINCT',
                          ct='TOTAL',
                          ptv='PTV')
```

FFG metrics

```{r}
ffgMet <- calcBentFFGmets(benthicsPrep,
                          masterTaxaListTarget,
                          sampID="BENSAMPID",
                          dist='IS_DISTINCT',
                          ct='TOTAL')
```

Habit metrics 
```{r}
habitMet <- calcBentHabitMets(benthicsPrep,
                          masterTaxaListTarget,
                          sampID="BENSAMPID",
                          dist='IS_DISTINCT',
                          ct='TOTAL',
                           habit = "HABIT")
```
Dominance metrics

```{r}
domMet <- calcBentDominMets(benthicsPrep,
                          masterTaxaListTarget,
                          sampID="BENSAMPID",
                          dist='IS_DISTINCT',
                          ct='TOTAL')

```


### Bonus special request metrics

These metrics are from the VDEQ biologists to supplement the aquametBio metrics for more regional/VA specific information.

```{r data prep}
source('aquametBio_specialFunctions/IBIspecialRequestFunctions_IsDistinct.R')

BCG <- read_excel('data/MASTER_ATTRIBUTES_BUGS_06062019.xlsx', sheet = 'Sheet1') %>% 
  dplyr::select(FinalID, GenusFinal, BCGattribute = `Virginia General BCG Attribute   First Choice`, 
                `BCG DisOxy`:`BCG pctIMP`) %>% 
  mutate_at(c("BCGattribute", "BCG DisOxy", "BCG acidity", "BCG Alkalinity", "BCG spCond", "BCG Chloride", 
              "BCG Sulfate", "BCG TN.TP",  "BCG totHab", "BCG RBS", "BCG pctIMP"), as.numeric) %>% 
  mutate(BCGattribute = case_when(BCGattribute == 0 ~ as.numeric(NA),
                                  TRUE ~ BCGattribute))

masterTaxaListBCG <- read_excel('data/masterTaxaGenus_01132022.xlsx', sheet = 'masterTaxaGenus') %>% # Emma update
  mutate_at(c("FamTolVal", "TolVal", "GVSCI_TolVal"), as.numeric) %>% 
  rename(FinalID_FFG = FFG) %>% # fix this before things get confusing
  
  
   # Optioservus fix
  mutate(Genus = case_when(FinalID == 'Promoresia' ~ 'Promoresia',
                           TRUE ~ as.character(Genus))) %>% 
  
   # Stenonema fix
  mutate(Genus = case_when(FinalID %in% c("Maccaffertium/Stenonema", "Maccaffertium") ~ "Maccaffertium",
                           TRUE ~ as.character(Genus)),
         GVSCI_TolVal = case_when(FinalID %in% c("Maccaffertium/Stenonema", "Maccaffertium") ~ 5.4,
                           TRUE ~ as.numeric(GVSCI_TolVal)) ) %>% #,
         # GregVSCI_TolVal2 = case_when(FinalID %in% c("Maccaffertium/Stenonema", "Maccaffertium") ~ 5.4,
         #                   TRUE ~ as.numeric(GregVSCI_TolVal2))) %>% 
         # 


  
  # add BCG information
  left_join(BCG, by = c('OldFinalID' = 'FinalID', 'GVSCI_FinalID' = 'GenusFinal')) %>% 
  
  
  # rename to desired format
  rename(PHYLUM = Phylum, 
         CLASS = Class,
         ORDER = Order, 
         FAMILY = Family,
         SUBFAMILY = Subfamily, 
         TRIBE = Tribe, 
         GENUS = Genus,
         # use genus level bugs for metrics
         #TARGET_TAXON = GVSCI_FinalID,
         FFG = GVSCI_FFG,
         HABIT = GVSCI_Habit, 
         PTV = GVSCI_TolVal) %>% 
  #mutate(TAXA_ID = 1:n()) %>%  # give numeric unique identifier
  mutate_if(is.character, str_to_upper) %>% # package needs uppercase taxa names
  
  # recode FFG to desired format: CF=collector-filterer, CG=collector-gatherer,PA=parasite,PI=Piercer,PR=predator,SC=scraper,SH=shredder
  mutate(FFG = case_when(FFG == "COLLECTOR" ~ "CG", 
                         FFG == "SCRAPER" ~ "SC",
                         FFG == "SHREDDER" ~ "SH",
                         FFG == "PREDATOR" ~ "PR",
                         FFG == "FILTERER" ~ "CF",
                         TRUE ~ as.character(NA)),
         #  recode HABIT to desired format:  AT=attached,BU=burrower,CB=climber, CN=clinger,DV=diver,PK=planktonic,SK=skater,SP=sprawler,SW=swimmer    
         HABIT = case_when(HABIT == "SWIMMER" ~ "SW", 
                           HABIT == "CLINGER" ~ "CN",
                           HABIT ==  "CLIMBER" ~ "CB",
                           HABIT == "BURROWER" ~ "BU",
                           HABIT == "SPRAWLER" ~ "SP",
                           TRUE ~ as.character(NA))) 
  

masterTaxaListBCG[ masterTaxaListBCG == "NA" ] <- NA

masterTaxaListBCGTarget <- masterTaxaListBCG %>% 
  #filter(Target == 1) # this will drop all higher level taxa from joining, not a good idea
  arrange(GVSCI_FinalID, Target) %>% 
  distinct(GVSCI_FinalID, .keep_all = T) %>% 
  rename(TARGET_TAXON = GVSCI_FinalID) %>% 
  mutate(TAXA_ID = 1:n()) %>% 
  as.data.frame()

bonusMetrics <- IBImetrics(benthicsPrep, masterTaxaListBCGTarget)
```

### Quick QA step: Tolerance value reevalutation

Greg suggested we may not have the HBI range we want with our current tolerance value spread (nothing less than 2) so Jason recalululated the "good" BCG tolerance value math to expand the top end (similar to how he had for the bad end) and that info is stored in the GregVSCI_TolVal2 column. 

For this quick test, we will reorganize the benthic data and rerun the tolerance metrics against that dataset. We will need to redo the math for the Tolerance metrics and the "bonus VA metrics" because some of the EPT taxa math includes tolerance thresholds for a few metrics.

```{r redo above with greg tol values}
masterTaxaList <- read_excel('data/masterTaxaGenus_01132022_JRH.xlsx',  sheet = 'masterTaxaGenusJRHFeb10') %>%  #change with genus fixes
  #read_excel('data/masterTaxaGenus_01132022.xlsx', sheet = 'masterTaxaGenus') %>% # Emma update
  mutate_at(c("FamTolVal", "TolVal", "GVSCI_TolVal", 'GregVSCI_TolVal2'), as.numeric) %>% 
  rename(FinalID_FFG = FFG) %>% # fix this before things get confusing
  
  # Optioservus fix
  mutate(Genus = case_when(FinalID == 'Promoresia' ~ 'Promoresia',
                           TRUE ~ as.character(Genus))) %>% 
  
   # Stenonema fix
  mutate(Genus = case_when(FinalID %in% c("Maccaffertium/Stenonema", "Maccaffertium") ~ "Maccaffertium",
                           TRUE ~ as.character(Genus)),
         # GVSCI_TolVal = case_when(FinalID %in% c("Maccaffertium/Stenonema", "Maccaffertium") ~ 5.4,
         #                   TRUE ~ as.numeric(GVSCI_TolVal)),
         GregVSCI_TolVal2 = case_when(FinalID %in% c("Maccaffertium/Stenonema", "Maccaffertium") ~ 5.4,
                           TRUE ~ as.numeric(GregVSCI_TolVal2))) %>% 
  

  
  # rename to desired format
  rename(PHYLUM = Phylum, 
         CLASS = Class,
         ORDER = Order, 
         FAMILY = Family,
         SUBFAMILY = Subfamily, 
         TRIBE = Tribe, 
         GENUS = Genus,
         # use genus level bugs for metrics
         #TARGET_TAXON = GVSCI_FinalID,
         FFG = GVSCI_FFG,
         HABIT = GVSCI_Habit, 
         
         ### Note change to greg tol value
         
         PTV = GregVSCI_TolVal2) %>% 
  #mutate(TAXA_ID = 1:n()) %>%  # give numeric unique identifier
  mutate_if(is.character, str_to_upper) %>% # package needs uppercase taxa names
  
  # recode FFG to desired format: CF=collector-filterer, CG=collector-gatherer,PA=parasite,PI=Piercer,PR=predator,SC=scraper,SH=shredder
  mutate(FFG = case_when(FFG == "COLLECTOR" ~ "CG", 
                         FFG == "SCRAPER" ~ "SC",
                         FFG == "SHREDDER" ~ "SH",
                         FFG == "PREDATOR" ~ "PR",
                         FFG == "FILTERER" ~ "CF",
                         TRUE ~ as.character(NA)),
 #  recode HABIT to desired format:  AT=attached,BU=burrower,CB=climber, CN=clinger,DV=diver,PK=planktonic,SK=skater,SP=sprawler,SW=swimmer    
         HABIT = case_when(HABIT == "SWIMMER" ~ "SW", 
                           HABIT == "CLINGER" ~ "CN",
                           HABIT ==  "CLIMBER" ~ "CB",
                           HABIT == "BURROWER" ~ "BU",
                           HABIT == "SPRAWLER" ~ "SP",
                           TRUE ~ as.character(NA))) 
    
masterTaxaList[ masterTaxaList == "NA" ] <- NA

masterTaxaListTarget <- masterTaxaList %>% 
  #filter(Target == 1) # this will drop all higher level taxa from joining, not a good idea
  arrange(GVSCI_FinalID, Target) %>% 
  
    filter(!is.na(GVSCI_FinalID)) %>% 

  
  distinct(GVSCI_FinalID, .keep_all = T) %>% 
  rename(TARGET_TAXON = GVSCI_FinalID) %>% 
  mutate(TAXA_ID = 1:n()) %>% 
  as.data.frame()
  #rowwise() %>% 
  #mutate(TAXA_ID = sum(3000, TAXA_ID))



benthics <- read_csv('data/aquametBioDataOut/finalBenthics.csv')
  #read.csv('data/aquametBioDataOut/benthics01132022.csv') 

benthics <- benthics %>% #[1:6921,] %>% 
  
  #filter(BenSampID == '3-RAP10109') %>% #'CUB7141') %>% #
  
  rename(BENSAMPID = BenSampID) %>%  # assignDistinct() changes names to caps so adjust here to avoid issues later
  mutate(FinalID = as.character(FinalID)) %>% 
  mutate_if(is.character, str_to_upper) %>%   # convert names to all caps
  
  ## join to full master taxa list
  #left_join(dplyr::select(masterTaxaList, FinalID, TARGET_TAXON, TAXA_ID),  by = "FinalID")
  left_join(dplyr::select(masterTaxaList, FinalID, GVSCI_FinalID#TARGET_TAXON),  by = "FinalID")
  ),  by = "FinalID") %>% 
  # join to target level master taxa list for unique TAXA_ID
  left_join(dplyr::select(masterTaxaListTarget, TARGET_TAXON, TAXA_ID),  by = c("GVSCI_FinalID" = 'TARGET_TAXON')) #"FinalID"
 # left_join(dplyr::select(masterTaxaListTarget, FinalID, TARGET_TAXON, TAXA_ID),  by = "FinalID")
  

  
benthicsPrep <- prepBentCts_WSA_EVJ(
  inCts = benthics,
  inTaxa = masterTaxaListTarget,
  sampID = "BENSAMPID",
  ct = "Individuals",
  taxa_id ="TAXA_ID"# "FinalID"#"TAXA_ID"
)


### aquametbio tol metrics with greg tol value
outTolGreg <- calcBentTolMets(benthicsPrep,
                          masterTaxaListTarget,
                          sampID="BENSAMPID",
                          dist='IS_DISTINCT',
                          ct='TOTAL',
                          ptv='PTV')

## va metrics with greg tol value
BCG <- read_excel('data/MASTER_ATTRIBUTES_BUGS_06062019.xlsx', sheet = 'Sheet1') %>% 
  dplyr::select(FinalID, GenusFinal, BCGattribute = `Virginia General BCG Attribute   First Choice`, 
                `BCG DisOxy`:`BCG pctIMP`) %>% 
  mutate_at(c("BCGattribute", "BCG DisOxy", "BCG acidity", "BCG Alkalinity", "BCG spCond", "BCG Chloride", 
              "BCG Sulfate", "BCG TN.TP",  "BCG totHab", "BCG RBS", "BCG pctIMP"), as.numeric) %>% 
  mutate(BCGattribute = case_when(BCGattribute == 0 ~ as.numeric(NA),
                                  TRUE ~ BCGattribute))

masterTaxaListBCG <- read_excel('data/masterTaxaGenus_01132022_JRH.xlsx',  sheet = 'masterTaxaGenusJRHFeb10') %>%  #change with genus fixes
  #read_excel('data/masterTaxaGenus_01132022.xlsx', sheet = 'masterTaxaGenus') %>% # Emma update
  mutate_at(c("FamTolVal", "TolVal", "GVSCI_TolVal", 'GregVSCI_TolVal2'), as.numeric) %>% 
  rename(FinalID_FFG = FFG) %>% # fix this before things get confusing
  
   # Optioservus fix
  mutate(Genus = case_when(FinalID == 'Promoresia' ~ 'Promoresia',
                           TRUE ~ as.character(Genus))) %>% 
  
   # Stenonema fix
  mutate(Genus = case_when(FinalID %in% c("Maccaffertium/Stenonema", "Maccaffertium") ~ "Maccaffertium",
                           TRUE ~ as.character(Genus)),
         # GVSCI_TolVal = case_when(FinalID %in% c("Maccaffertium/Stenonema", "Maccaffertium") ~ 5.4,
         #                   TRUE ~ as.numeric(GVSCI_TolVal)),
         GregVSCI_TolVal2 = case_when(FinalID %in% c("Maccaffertium/Stenonema", "Maccaffertium") ~ 5.4,
                           TRUE ~ as.numeric(GregVSCI_TolVal2))) %>% 

  
  # add BCG information
  left_join(BCG, by = c('OldFinalID' = 'FinalID', 'GVSCI_FinalID' = 'GenusFinal')) %>% 
  
  
  # rename to desired format
  rename(PHYLUM = Phylum, 
         CLASS = Class,
         ORDER = Order, 
         FAMILY = Family,
         SUBFAMILY = Subfamily, 
         TRIBE = Tribe, 
         GENUS = Genus,
         # use genus level bugs for metrics
         #TARGET_TAXON = GVSCI_FinalID,
         FFG = GVSCI_FFG,
         HABIT = GVSCI_Habit, 
         
         
         
         ### Note change to greg tol value
         
         PTV = GregVSCI_TolVal2) %>% 
  #mutate(TAXA_ID = 1:n()) %>%  # give numeric unique identifier
  mutate_if(is.character, str_to_upper) %>% # package needs uppercase taxa names
  
  # recode FFG to desired format: CF=collector-filterer, CG=collector-gatherer,PA=parasite,PI=Piercer,PR=predator,SC=scraper,SH=shredder
  mutate(FFG = case_when(FFG == "COLLECTOR" ~ "CG", 
                         FFG == "SCRAPER" ~ "SC",
                         FFG == "SHREDDER" ~ "SH",
                         FFG == "PREDATOR" ~ "PR",
                         FFG == "FILTERER" ~ "CF",
                         TRUE ~ as.character(NA)),
         #  recode HABIT to desired format:  AT=attached,BU=burrower,CB=climber, CN=clinger,DV=diver,PK=planktonic,SK=skater,SP=sprawler,SW=swimmer    
         HABIT = case_when(HABIT == "SWIMMER" ~ "SW", 
                           HABIT == "CLINGER" ~ "CN",
                           HABIT ==  "CLIMBER" ~ "CB",
                           HABIT == "BURROWER" ~ "BU",
                           HABIT == "SPRAWLER" ~ "SP",
                           TRUE ~ as.character(NA))) 
  

masterTaxaListBCG[ masterTaxaListBCG == "NA" ] <- NA

masterTaxaListBCGTarget <- masterTaxaListBCG %>% 
  #filter(Target == 1) # this will drop all higher level taxa from joining, not a good idea
  arrange(GVSCI_FinalID, Target) %>% 
  distinct(GVSCI_FinalID, .keep_all = T) %>% 
  rename(TARGET_TAXON = GVSCI_FinalID) %>% 
  mutate(TAXA_ID = 1:n()) %>% 
  as.data.frame()

bonusMetricsGreg <- IBImetrics(benthicsPrep, masterTaxaListBCGTarget)



```


Some quick mutations to make comparison easier

```{r greg mutations}
tolCompare <- bind_rows(outTol %>% mutate(Source = 'Original') %>% dplyr::select(Source, everything()),
                        outTolGreg %>% mutate(Source = 'Greg') %>% dplyr::select(Source, everything())) %>% 
  arrange(BENSAMPID, Source)

vaCompare <- bind_rows(bonusMetrics %>% mutate(Source = 'Original') %>% dplyr::select(Source, everything()),
                       bonusMetricsGreg %>% mutate(Source = 'Greg') %>% dplyr::select(Source, everything())) %>% 
  arrange(BENSAMPID, Source)
```



### Add sheet with reference/stress site info to final dataset

For model development

```{r ref/stress info}
refstress <- read_excel('data/RefSitesSourceGIS.xlsx', sheet = 'AllBugGenus') %>% 
  mutate(JRH_Calls = case_when(JRH_Calls ==  "Ref-Pied" ~ "Ref",
                               TRUE ~ JRH_Calls)) # collapse piedmont ref to ref for this exercise
```



## for review


Write .xlsx with different sheets for types of metrics

```{r}
library(openxlsx)
outBensamps <- benSamps %>% dplyr::select(StationID, BenSampID , RepNum, `Collection Date`, `Sample Comments`,
                                          `Collected By`, Taxonomist, Gradient, `Target Count`, Season, everything()) %>% 
  left_join(refstress %>% 
              dplyr::select(StationID, RefStress = JRH_Calls) %>% 
              distinct(StationID, .keep_all = T), by = 'StationID' ) %>% 
  dplyr::select(DataSource, RefStress, everything())

list_of_datasets <- list("BenthicsRarified" = benthics,
                         "BenSamps" = outBensamps, 
                         "Taxa Metrics" = outTax, 
                         "Tolerance Metrics" = outTol, 
                         "Tolerance Metrics_GREG" = outTolGreg, 
                         "Tolerance Metrics Comparison" = tolCompare,
                         "FFG Metrics" = ffgMet, 
                         "Habit Metrics" = habitMet,
                         "Dominance Metrics" = domMet,
                         "VA Requested Metrics" = bonusMetrics,
                         "VA Requested Metrics_GREG" = bonusMetricsGreg,
                         "VA Requested metrics Comparison" = vaCompare)
write.xlsx(list_of_datasets, file = "data/aquametBioDataOut/aquametBio_final_02162022.xlsx")

```


## for model development 
Organize VA bonus metrics into appropriate metric categories

```{r}
# get one object to select from to make life easier. paste greg in front of metrics that use that tol value scale
allMetrics <- left_join(outTax, outTol, by = 'BENSAMPID') %>% 
  left_join(ffgMet, by = 'BENSAMPID') %>%
  left_join(outTolGreg %>% setNames(paste0('gregTol_', names(.))), by = c('BENSAMPID' = 'gregTol_BENSAMPID')) %>%
  left_join(habitMet, by = 'BENSAMPID') %>%
  left_join(domMet, by = 'BENSAMPID') %>%
  left_join(bonusMetrics, by = 'BENSAMPID') %>% 
  left_join(bonusMetricsGreg %>% setNames(paste0('gregTol_', names(.))), by = c('BENSAMPID' = 'gregTol_BENSAMPID')) 

# slim list for easy sorting
finalMetrics <- outTax %>% 
  left_join(bonusMetrics, by = 'BENSAMPID') %>% 
  left_join(bonusMetricsGreg %>% setNames(paste0('gregTol_', names(.))), by = c('BENSAMPID' = 'gregTol_BENSAMPID')) 

composition <- dplyr::select(finalMetrics, BENSAMPID, 
                             AMPHPIND, AMPHPTAX, CHIRPIND, CHIRPTAX, CRUSPIND, CRUSPTAX, DIPTPIND, DIPTPTAX, EPHEPIND, EPHEPTAX,
                             EPOTPIND, EPOTPTAX, EPT_PIND, EPT_PTAX, HEMIPIND, HEMIPTAX, MITEPIND, MITEPTAX, MOLLPIND, MOLLPTAX,
                             NOINPIND, NOINPTAX, ODONPIND, ODONPTAX, OLLEPIND, OLLEPTAX, ORTHPIND, ORTHPTAX, PLECPIND, PLECPTAX,
                             TANYPIND, TANYPTAX, TRICPIND, TRICPTAX, TUBINAIDPIND, TUBINAIDPTAX,
                             `%Ephem`, `%PT - Hydropsychidae`,  `%GenusScraper`, `%Chiro`, `%EPT-H+C`, `%Ephem-B`,
                             `%EPT4.5`,	`%EPT6.5`,
                             `gregTol_%Ephem`, `gregTol_%PT - Hydropsychidae`, `gregTol_%GenusScraper`, `gregTol_%Chiro`, 
                             `gregTol_%EPT-H+C`, `gregTol_%Ephem-B`,  `gregTol_%EPT4.5`,	`gregTol_%EPT6.5`)
richness <- dplyr::select(finalMetrics, BENSAMPID,
                          TOTLNTAX,	AMPHNTAX,	CHIRNTAX,	CRUSNTAX,	DIPTNTAX,	EPHENTAX,	EPOTNTAX,	EPT_NTAX,	HEMINTAX,	MITENTAX,	
                          MOLLNTAX,	NOINNTAX,	ODONNTAX,	OLLENTAX,	ORTHNTAX,	PLECNTAX,	TANYNTAX,	TRICNTAX,	TUBINAIDNTAX,
                          `Genus Total Taxa`, `Genus EPT Taxa`, Elmid,
                          `gregTol_Genus Total Taxa`, `gregTol_Genus EPT Taxa`, gregTol_Elmid)
tolerance <- left_join(outTol, outTolGreg %>% setNames(paste0('gregTol_', names(.))), by = c('BENSAMPID' = 'gregTol_BENSAMPID')) %>% 
  left_join(bonusMetrics %>% 
              dplyr::select(BENSAMPID, `Genus HBI`, 	`BCGatt2&3`,	`%BCGatt2&3`,	BCGatt5,	`%BCGatt5`,	`%BCG_DO_att2&3`,	`%BCG_DO_att5`,
                            `%BCG_Acidity_att2&3`,	`%BCG_Acidity_att5`,	`%BCG_Alkalinity_att2&3`,	`%BCG_Alkalinity_att5`,	`%BCG_spCond_att2&3`,
                            `%BCG_spCond_att5`,	`%BCG_Chloride_att2&3`,	`%BCG_Chloride_att5`,	`%BCG_Sulfate_att2&3`,	`%BCG_Sulfate_att5`,
                            `%BCG_TN.TP_att2&3`,	`%BCG_TN.TP_att5`,	`%BCG_totHab_att2&3`,	`%BCG_totHab_att5`,	`%BCG_RBS_att2&3`,
                            `%BCG_RBS_att5`,	`%BCG_pctIMP_att2&3`,	`%BCG_pctIMP_att5`), by = 'BENSAMPID') %>% 
  left_join(bonusMetricsGreg %>% 
              setNames(paste0('gregTol_', names(.))) %>% 
              dplyr::select(gregTol_BENSAMPID, `gregTol_Genus HBI`, `gregTol_BCGatt2&3`,
                            `gregTol_%BCGatt2&3`,	gregTol_BCGatt5,	`gregTol_%BCGatt5`,	`gregTol_%BCG_DO_att2&3`,	`gregTol_%BCG_DO_att5`,
                            `gregTol_%BCG_Acidity_att2&3`,	`gregTol_%BCG_Acidity_att5`,	`gregTol_%BCG_Alkalinity_att2&3`,	
                            `gregTol_%BCG_Alkalinity_att5`,	`gregTol_%BCG_spCond_att2&3`, `gregTol_%BCG_spCond_att5`,
                            `gregTol_%BCG_Chloride_att2&3`,	`gregTol_%BCG_Chloride_att5`,	`gregTol_%BCG_Sulfate_att2&3`,	`gregTol_%BCG_Sulfate_att5`,
                            `gregTol_%BCG_TN.TP_att2&3`,	`gregTol_%BCG_TN.TP_att5`,	`gregTol_%BCG_totHab_att2&3`,	`gregTol_%BCG_totHab_att5`,
                            `gregTol_%BCG_RBS_att2&3`, `gregTol_%BCG_RBS_att5`,	`gregTol_%BCG_pctIMP_att2&3`,	`gregTol_%BCG_pctIMP_att5`),
            by = c('BENSAMPID' = 'gregTol_BENSAMPID')) 
dominance <- left_join(domMet, bonusMetrics %>% dplyr::select(BENSAMPID, `Genus %2 Dominant`), by = 'BENSAMPID') %>% 
  left_join(bonusMetricsGreg %>% 
              setNames(paste0('gregTol_', names(.))) %>% 
              dplyr::select(gregTol_BENSAMPID, `gregTol_Genus %2 Dominant`), by = c('BENSAMPID' = 'gregTol_BENSAMPID')) 
ffg <- left_join(ffgMet, bonusMetrics %>% 
                   dplyr::select(BENSAMPID, `%Collector` ), by = 'BENSAMPID') %>% 
  left_join(bonusMetricsGreg %>% 
              setNames(paste0('gregTol_', names(.))) %>% 
              dplyr::select(gregTol_BENSAMPID, `gregTol_%Collector`), by =  c('BENSAMPID' = 'gregTol_BENSAMPID'))
habit <- left_join(habitMet, bonusMetrics %>% 
                   dplyr::select(BENSAMPID, `%Clinger-HS`), by = 'BENSAMPID') %>% 
  left_join(bonusMetricsGreg %>% 
              setNames(paste0('gregTol_', names(.))) %>% 
              dplyr::select(gregTol_BENSAMPID, `gregTol_%Clinger-HS` ), by =  c('BENSAMPID' = 'gregTol_BENSAMPID'))


# make sure no metrics left behind
outMets <- left_join(composition, richness, by = 'BENSAMPID') %>% 
  left_join(tolerance, by = 'BENSAMPID') %>% 
  left_join(dominance, by = 'BENSAMPID') %>% 
  left_join(habit, by = 'BENSAMPID') %>% 
  left_join(ffg, by = 'BENSAMPID') 

 names(allMetrics)[! names(allMetrics) %in% names(outMets)]
#"TOTLNIND"  only metric left and that's fine to drop!


library(openxlsx)
outBensamps <- benSamps %>% dplyr::select(StationID, BenSampID , RepNum, `Collection Date`, `Sample Comments`,
                                          `Collected By`, Taxonomist, Gradient, `Target Count`, Season, everything()) %>% 
  left_join(refstress %>% 
              dplyr::select(StationID, RefStress = JRH_Calls) %>% 
              distinct(StationID, .keep_all = T), by = 'StationID' ) %>% 
  dplyr::select(DataSource, RefStress, everything())

list_of_datasets <- list("BenSamps" = outBensamps, 
                         "Composition" = composition, 
                         "Richness" = richness, 
                         "Tolerance" = tolerance, 
                         "Dominance" = dominance,
                         "Habit" = habit, 
                         "FFG" = ffg)
write.xlsx(list_of_datasets, file = "data/aquametBioDataOut/aquametBio_final_02222022.xlsx")

```






plot greg vs reg tol break

```{r}

compare <- left_join(vaCompare, outBensamps, by = c('BENSAMPID' = 'BenSampID')) %>% 
  mutate(Source = as.factor(Source)) 
  
ggplot(compare, aes(x = RefStress, y = `Genus HBI` , colour = Source ) )+
   geom_boxplot()

ggplot(compare %>% filter(RefStress %in% c('Ref', "Str")), aes(x = RefStress, y = `Genus HBI` , colour = Source ) )+
   geom_boxplot()
```


More interactive version


```{r}
library(esquisse)
esquisser(compare)
```
```{r}
# library(tidyverse)
# library(readxl)
# library(esquisse)
# vaCompare <- read_excel("data/aquametBioDataOut/aquametBio_final_02162022.xlsx", sheet = 'VA Requested metrics Comparison')
# outBensamps<- read_excel("data/aquametBioDataOut/aquametBio_final_02162022.xlsx", sheet = 'BenSamps')
# compare <- left_join(vaCompare, outBensamps, by = c('BENSAMPID' = 'BenSampID')) %>%
#   mutate(Source = as.factor(Source))
# 
# esquisser(compare)

```


How many samples with <50 bugs, > 220, etc

```{r understand data} 
test <- left_join(outTax, outBensamps, by = c('BENSAMPID' = 'BenSampID')) %>% 
  #filter(RefStress == 'Ref') %>% 
  filter(TOTLNIND > 220) %>% 
  dplyr::select(StationID, `Collection Date`, everything())

```
